Move mission control viewmodels into missioncontrol folder
