+++ src/GRA.Controllers/ViewModel/MissionControl/Challenges/ChallengesListViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
 using System.Collections.Generic;
 
-namespace GRA.Controllers.ViewModel.Challenges
+namespace GRA.Controllers.ViewModel.MissionControl.Challenges
 {
     public class ChallengesListViewModel
     {
