+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/BookAddViewModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace GRA.Controllers.ViewModel.Participants
+namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
     public class BookAddViewModel
     {
