+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/BookListViewModel.cs
@@ -1,16 +1,16 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
 using System.Collections.Generic;
 
-namespace GRA.Controllers.ViewModel.Participants
+namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
-    public class HistoryListViewModel
+    public class BookListViewModel
     {
-        public IEnumerable<GRA.Domain.Model.UserLog> Historys { get; set; }
+        public List<GRA.Domain.Model.Book> Books { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public int Id { get; set; }
         public int HouseholdCount { get; set; }
         public int? HeadOfHouseholdId { get; set; }
         public bool HasAccount { get; set; }
-        public bool CanRemoveHistory { get; set; }
+        public bool CanModifyBooks { get; set; }
     }
 }
