+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/HistoryListViewModel.cs
@@ -1,16 +1,16 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
 using System.Collections.Generic;
 
-namespace GRA.Controllers.ViewModel.Participants
+namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
-    public class MailListViewModel
+    public class HistoryListViewModel
     {
-        public IEnumerable<GRA.Domain.Model.Mail> Mails { get; set; }
+        public IEnumerable<GRA.Domain.Model.UserLog> Historys { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public int Id { get; set; }
         public int HouseholdCount { get; set; }
         public int? HeadOfHouseholdId { get; set; }
         public bool HasAccount { get; set; }
-        public bool CanRemoveMail { get; set; }
+        public bool CanRemoveHistory { get; set; }
     }
 }
