+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/MailDetailViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace GRA.Controllers.ViewModel.Participants
+﻿namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
     public class MailDetailViewModel
     {
