+++ src/GRA.Web/Areas/MissionControl/Views/Challenges/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Challenges.ChallengesDetailViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Challenges.ChallengesDetailViewModel
 
 <form asp-controller="Challenges" method="post" role="form">
