+++ src/GRA.Web/Areas/MissionControl/Views/Challenges/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Challenges.ChallengesListViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Challenges.ChallengesListViewModel
 
 <div class="row">
     <div class="modal fade" id="deleteModal" tabindex="-1" role="dialog" aria-labelledby="deleteModalLabel">
