+++ src/GRA.Web/Areas/MissionControl/Views/Participants/AddBook.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.BookAddViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.BookAddViewModel
 
 <form asp-controller="Participants" asp-action="AddBook" asp-route-id="@Model.Id" method="post" role="form">
     <div class="row row-spacing">
