+++ src/GRA.Web/Areas/MissionControl/Views/Participants/Books.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.BookListViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.BookListViewModel
 
 @Html.Partial("_ParticipantPartial")
