+++ src/GRA.Web/Areas/MissionControl/Views/Participants/Detail.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.ParticipantsDetailViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.ParticipantsDetailViewModel
 
 @Html.Partial("_ParticipantPartial")
