+++ src/GRA.Web/Areas/MissionControl/Views/Participants/History.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.HistoryListViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.HistoryListViewModel
 
 @Html.Partial("_ParticipantPartial")
