+++ src/GRA.Web/Areas/MissionControl/Views/Participants/Household.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.HouseholdListViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.HouseholdListViewModel
 
 @Html.Partial("_ParticipantPartial")
