+++ src/GRA.Web/Areas/MissionControl/Views/Participants/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.ParticipantsListViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.ParticipantsListViewModel
 
 
 @if (Model.CanRemoveParticipant)
