+++ src/GRA.Web/Areas/MissionControl/Views/Participants/Mail.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.MailListViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.MailListViewModel
 
 @Html.Partial("_ParticipantPartial")
