+++ src/GRA.Web/Areas/MissionControl/Views/Participants/MailDetail.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.MailDetailViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.MailDetailViewModel
 
 @if (Model.CanRemoveMail)
 {
