+++ src/GRA.Web/Areas/MissionControl/Views/Participants/PasswordReset.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Participants.PasswordResetViewModel
+﻿@model GRA.Controllers.ViewModel.MissionControl.Participants.PasswordResetViewModel
 
 @Html.Partial("_ParticipantPartial")
